добавил интеграционные тесты, проходят но нужно рефачить

+++ src/Core/OrderPaymentSystem.Application/DTOs/Product/ProductDto.cs
@@ -1,3 +1,3 @@
 ﻿namespace OrderPaymentSystem.Application.DTOs.Product;
 
-public record ProductDto(long Id, string Name, string Description, decimal Price, int StockQuantity, DateTime CreatedAt);
+public record ProductDto(int Id, string Name, string Description, decimal Price, int StockQuantity, DateTime CreatedAt);
